add execute file

+++ DDL/sql_files/article_likes.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `article_likes` (
-  `id` bigint NOT NULL AUTO_INCREMENT,
-  `user_id` bigint NOT NULL,
-  `article_id` bigint NOT NULL,
+  `id` int NOT NULL AUTO_INCREMENT,
+  `user_id` int NOT NULL,
+  `article_id` int NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE = InnoDB DEFAULT CHARSET = utf8;
